build luajit both with and wthout lua 5.2 compat changes
Luajit is first and foremost implementation of Lua 5.1 hence we should
aim for best compatibility with Lua 5.1. It is commonly known fact that
Lua versions are not backward compatible so ie Lua 5.2 breaks
compatibility with 5.1, 5.3 breaks compatibility with 5.2 and so does
5.4 with 5.3. Luajit already enables some of Lua 5.2 extensions by
default, but not those that introduce incompatibilities with Lua 5.1.
The latter are ebabled with LUAJIT_ENABLE_LUA52COMPAT.

Also Luajit no longer loads system Lua 5.1 modules (common between all
Lua 5.1 interpreters) which it did since forever and its generally
easier to maintain dependcies if it's known it continues to do so.

Hence build Luajit with 5.2 extensions for packages that need it, but
most importantly keep building Luajit without them for others.